requirements_docker.txt
Flask==3.1.0
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
psutil==5.9.8
psycopg2-binary==2.9.7
